// frontend/src/app/models/user.ts
export default class User {
  korisnickoIme: string = "";
  lozinka: string = "";
  ime: string = "";
  prezime: string = "";
  tip: string = "";
  pol: string = "";
  adresa: string = "";
  brojTelefona: string = "";
  email: string = "";
  brojKreditneKartice: string = "";
  profilnaSlika: string = "";
  blokiran: number = 0;
  firma: string = ""
}